requirements.txt
streamlit==0.63.0
pandas==1.0.5
numpy==1.16.6
biopython==1.76
matplotlib==3.2.2
requests==2.24.0